+++ examples/.prettierrc.js
@@ -4,4 +4,5 @@ module.exports = {
     jsxBracketSameLine: true,
     singleQuote: true,
     trailingComma: 'all',
+    arrowParens: 'avoid',
 };